Added redirect handling such that the original URLs are stored in a table so that when lookups are later done with those URLs, we can find the arrived at URL to look up the page by.
Added a service and controller method to get the list of fetched at
timestamps for an URL. This powers a new dropdown that allows us to
see older versions of the page.

Added snapshot URL not found handling so the errors are more clear
if the user tries to take a snapshot of a page that is returning
an HTTP error code.